cli: add new `dev` command to run functions locally simulating production experience
The implementation is initial but implements a minimal simulation to run Fleet functions locally, all Node.js APIs are available locally but we want to disable some APIs or provide warnings if used to be more faithful to the production environment.

This implementation watches at the `fleet.yml` configuration files and uses the Webpack to compile and watch at the functions.

- When the `fleet.yml` file is modified the watcher will be restarted
- Modifying any function will make a new compilation
- Simulates the Response and Request APIs
- Dynamic paths work the same as in production

This implementation ignores asynchronous limits and concurrency because it is not necessary for a development environment, it is related to scale in production.

We also don't add a memory limit or throttle for CPU, because they're more difficult to simulate, maybe we can explore limiting memory of running the functions in a worker thread.