cli: Add OOTB TypeScript support and big build performance improvements
Well, the goal is to provide OOTB TypeScript support without having to add extra configuration to get started. The implementation is also not opinionated about the settings so we use `@babel/preset-typescript` which does not do type checking at build time and will not degrade build time performance.

As part of the support also adding some cache level settings for webpack and for babel itself, this brings considerably decreases the build time for larger projects and medium projects, decreasing around 11x in some cases 🎉.