cli: adds framework support for dev command
Framework support allows Fleet CLI to bypass the build with the webpack and use the build that was made from the framework to start the development environment.

In this first interaction, we only support our framework as a validation test, this means that the build is made by the framework and Fleet CLI constantly watches the changes in `fleet.yml` to define the routes.